Findbugs fixes, some other tidying up.
